MAINT: add a patch for `CartPoleEnv.__init__`
In `CartPoleEnv.__init__`, function call of `spaces.Box()` does not
explicitly assign a type to `dtype`.
In currently version of gym (0.10.5), it leads a warning being
issued when `CartPoleEnv` is created. (more precisely, any calling
in other environments will make it happen, too.)

Therefore, it is patched through modifying the bytecode in order
that the warning won't be issued anymore even if there is a child
class calling `CartPoleEnv.__init__()`.